[Feat] 게임 초대 응답 api 코드 작성 #75
